HotPAgent/requirements.txt
playwright==1.41.0
selenium==4.17.0
langchain==0.1.4
langchain-openai==0.0.5
pandas==2.2.0
requests==2.31.0
beautifulsoup4==4.12.3
aiohttp==3.9.3
python-dotenv==1.0.0
pydantic==2.5.3
schedule==1.2.0
loguru==0.7.2
